+++ react-app/src/components/Main/ChannelBar/EditChannel/index.js
@@ -1,0 +1,10 @@
+const EditChannel = ({ setShowModal }) => {
+	return (
+		<div className="edit-channel-icon" onClick={() => setShowModal(true)}>
+			<div>Edit Channel</div>
+			<i className="fa-solid fa-gear"></i>
+		</div>
+	);
+};
+
+export default EditChannel;
